Fix duplicate time in the schedule hang
This fixes issue #178, which happens if a duplicate of the first entry
in the list appears.  In this case an infinite loop occurs because the
list ends up being circler.  In all cases a duplicate entry at the same
time in the list is wrong because 2 instructions are valid at the same
time.

This fix causes a duplicate entry to result in a failure of accepting
the schedule.